Move global timezone methods into ApplicationController
These are only called from the view anyway
